added new branch and new counter
