--- src/Contents/Counter/ButtonCount/ButtonCount.module.css
@@ -1,8 +1,0 @@
-.buttons {
-    display: flex;
-    justify-content: center;
-}
-
-.button {
-    margin: 12px;
-}
