+++ src/Contents/Counter/CounterSet/CounterSet.tsx
@@ -1,7 +1,5 @@
 import React, {useState} from "react";
 import s from "./CounterSet.module.css";
-import InputCount from "../InputCount/InputCount";
-import ButtonCount from "../ButtonCount/ButtonCount";
 
 type CounterSetTypeProps = {
 
@@ -12,7 +10,7 @@ function CounterSet (props: CounterSetTypeProps) {
 
     return (
         <div className={s.counterUI}>
-            <div className={s.input1}>
+{/*            <div className={s.input1}>
                 <InputCount count={count}/>
             </div>
             <div className={s.input2}>
@@ -20,7 +18,7 @@ function CounterSet (props: CounterSetTypeProps) {
             </div>
             <div className={s.button}>
                 <ButtonCount count={count} setCount={setCount}/>
-            </div>
+            </div>*/}
         </div>
     );
 }
