+++ src/Contents/Counter/CounterUI/CounterUI.module.css
@@ -1,16 +1,33 @@
 .counterUI {
     display: grid;
-    grid-template-areas: "i i" "b b";
-    grid-template-rows: 2fr 1fr;
-    grid-template-columns: 1fr 1fr;
+    grid-template-areas: "i" "b";
+    grid-template-rows: 1fr 1fr;
     grid-gap: 15px;
 }
 
-.input {
-    margin-top: 10px;
+.inputs {
     grid-area: i;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 10px;
+}
+
+.input {
+    width: 150px;
+    height: 50px;
+    text-align: center;
+    font-size: 35px;
+}
+
+.buttons {
+    display: flex;
+    justify-content: center;
 }
 
 .button {
     grid-area: b;
+    margin: 12px;
+    width: 50px;
+    height: 30px;
 }
