+++ src/Contents/Counter/InputCount/InputCount.module.css
@@ -4,3 +4,11 @@
     align-items: center;
     padding: 10px;
 }
+
+.error {
+    border: red 1px solid;
+}
+
+.error-message {
+    color: red;
+}
